Remove useless ParameterMarkerExpressionSegmentBinder and extract new SubqueryTableBindUtils
